Computing Cook's distance.
Notice that the current implementation does not check whether the model becomes rank deficient when restricting to a certain subgroup. Whenever this occurs the regression cannot be performed, and since the implementation does not check for this Cortana will crash. Fixing this is next on my todo-list. On the plus side: in my tests with three datasets this problem occurred only once, so YMMV.

git-svn-id: https://svn.science.uu.nl/repos/sci.knobb101.SubgroupDiscovery@813 3a85a776-1598-11dc-ad04-adf9e166afdd