ConvexHullROC code implemented.
Seems correct, but another bug crept up.
SubgroupSet temporarily creates an ROCList and a ConvexHullROC.
The outputs differ, as ROCList does not deal with multiple Subgroups falling onto the same point.
This was not a problem in Safarii, as ROCList was only used to draw points.
In Cortana, the points on the hull are listed above the plot, and is thus potentially incomplete.

ConvexHullROC code should be thread safe, but needs to be tested.



git-svn-id: https://svn.science.uu.nl/repos/sci.knobb101.SubgroupDiscovery@2024 3a85a776-1598-11dc-ad04-adf9e166afdd