Candidate.CandidateComparatorDepthFirst update.
DEPTH_FIRST is not truly depth first.
For any RefinementList, all Refinements for the current Candidate/ ConditionList are evaluated in a breadth first fashion.
These Refinements are then added to the CandidateQueue.
From this queue the first Candidate is taken to refine further.

This posed a problem on depth=0, where the empty Subgroup is the only Candidate.
All Refinements are therefore done in the exact same way as in the BREADTH_FIRST setting.
To 'fake' / enforce DEPTH_FIRST like behaviour, these Refinements are put into the queue in a 'ConditionList=based' ordering.
This ensures that the Conditions, formed on d=0, are refined in the order of their creation on d=1.

(NOTE this is very similar to the once implemented natural and search order of ConditionList, but that code is no more/ need update.)



git-svn-id: https://svn.science.uu.nl/repos/sci.knobb101.SubgroupDiscovery@2089 3a85a776-1598-11dc-ad04-adf9e166afdd