SubgroupDiscovery.Filter and DebugFilter.
Remove useless Refinements from RefinementList.
Split into a PRE and POST stage.
PRE can be done on Condition.Operator only, so without 'instantiated values'.
POST can only be performed taking into account Condition.values.
Still both are in SD now, for uniform code, this may be subject to change in some future.

PRE - EQUALS
For any ConditionList containing (C=v), any Refinement on C is useless.
This is true for NOMINAL, NUMERIC and BINARY Columns, and for every SearchSetting and NominalOperatorSetting.

PRE - DEPTH_FIRST
For DEPTH_FIRST the number of Candidates to be valuated can be cut in half.
For any Condition, only Refinements with Conditions about later Column/Operator/value combinations need to evaluated.
O(n^2) -> O((n*(n+1)) / 2), triangulat, still n^2, but saves a lot of time.

POST - contains
There is no use in adding a Condition to a ConditionList that already exist in said list.

POST - BREADTH_FIRST create numeric equals
Any Refinement that would create an 'equals through LEQ and GEQ' is useless.
[ (C<=v) ^ (C>=v) ^ ...] is equivalent to [ (C=v) ^ ...].
For BREADTH_FIRST, NumericStrategy.NUMERIC_ALL, NumericOperatorSetting.NUMERIC_ALL (C=v) will be guaranteed to be evaluated on depth d.
Any Refinement that would create an 'equals through LEQ and GEQ' on depth d+1 is therefore useless.
Note for BINS it is not guaranteed that (C=v) is evaluated, as v relies on the 'reselection' of the unique numeric domain by the Subgroup.members.
Note for BEST it is guaranteed that (C=v) is evaluated, but it may not be used as seed for further refinement as it is not the BEST.
Note for no other setting than NominalOperatorSetting.Numeric_ALL (EQUALS, LEQ, GEQ) is this relevant.
Note similar reasoning holds for DEPTH_FIRST to, but this is taken care of the PRE-EQUALS test.

POST - has equals through LEQ and GEQ
Similar to PRE_EQUALS, but based on current ConditionList and to be added Condition for Refinement.
If LEQ and GEQ are valid NumericOperators, any addition of a Condition about C to a ConditionList that contains [ (C>=v) ^ (C<=v) ^ ...] is useless.
NOTE This should actually be a PRE check...

Filter is applied on RefinementList as a post-processing step, so after it is created.
Class and methods are in testing phase, code will be cleanup up later.

Summary of why Filter can save a lot of memcpy / GC cycles.
// for now the code creates a Condition that may not be needed at all
// the final code can be based on isUseful(Refinement, aValue)
// using Refinement.Subgroup.ConditionList, Refinement.ConditionBase, and aValue
// but at least Refinement.getRefinedSubgroup(aValue) is not called (expensive)
// it creates a copy of the Subgroup (members-bitset) and ConditionList
// and performs a full table scan to add the new Condition
// as the new Subgroup would not decrease in size / or became empty it was not
// logged as a Candidate (though it does increment itsNrCandidates)



git-svn-id: https://svn.science.uu.nl/repos/sci.knobb101.SubgroupDiscovery@2092 3a85a776-1598-11dc-ad04-adf9e166afdd