CandidateQueue.postProcessCBBS() Re-implementation of CBBS post-processing strategy.
First, covercount[nr_rows] keeps track of the cover count of each row.
It is updated when a Candidate is added to the beam.

Second, main loop does not consider every Candidate all the time.
The score a Candidate can achieve is upper-bounded by the score it achieved in the previous round.
So for Candidates still under consideration, if their lastScore is less than the current maxScore,
their new multiplicative weight score can never exceed the current maxScore either.
This avoids a lot of calls to the expensive computeMultiplicativeWeight().

Third, as the beam is empty on the first round, the weight for each Candidate is 1.
As a result, the highest scoring Candidate will always win the first round.
This part is taken out of the main loop, and scores are set directly.



git-svn-id: https://svn.science.uu.nl/repos/sci.knobb101.SubgroupDiscovery@2162 3a85a776-1598-11dc-ad04-adf9e166afdd