ConditionListBuilder will completely replace replace ConditionList. ConditionList extends ArrayList<Condition>, making it slow and unsafe. It can be manipulated in unwanted ways (eg. clear() while is should hold a Subgroups description).
ConditionListBuider creates immutable ConditionLists.
These are implemented internally in the ConditionListBuilder class.
No implementations are exposed, and the implementation classes can not be extended.
This ensures immutability and required behaviour.
External code need only call ConditionList size(), get(), compareTo() and toString().
All implementation classes are forced to implement these.
There are optimised implementation classes for 0, 1, 2 and 3 Condition ConditionLists.
An optimised version for 4-16 conjunct will follow in due time.

The Conditions in these ConditionLists are internally ordered in canonical order.
This makes comparison of ConditionLists fast and simple.
Redundancy is reduced.
Concurrency is improved.
Memory comsuption is reduced (ConditionLists are immutable, they are not copied, but shared as is).

toString() prints the conjunction in search order.

toCanonicalString() (to be implemented) will print the conjunction in canonical order.
(This allows for invocation invariant results in experiment using concurrency.)

WARNING
Old results are incomparable to new results, as (A ^ B) and (B ^ A) should no longer occur in result lists.
Checking for integrity is very hard, and needs special code.
(This is not implemented, as there has never been a rgression test suit.)
Therefore, if this code is incorrect, a massive amout of classes will have to be reset.



git-svn-id: https://svn.science.uu.nl/repos/sci.knobb101.SubgroupDiscovery@2176 3a85a776-1598-11dc-ad04-adf9e166afdd