### commit hang 24.10.2013 - this is a retry ###
ConditionBuilder will replace Condition class.
Internally, it holds a number of Condition implementation classes.
The old Condition class held 5 value member fields, and an Operator field.
This made the Condition objects both big, and unsafe/ cumbersome to maintain.
Each new implementation class holds only the value that corresponds to that Condition type.
No classes are exposed to the outside API, so any implementation can be updated freely.

NOTE this class is not used yet, as compareTo(Condition) needs to be implemented first.
evaluate() may not be implemented, as it is deprecated, slow, legacy code to be removed anyway.



git-svn-id: https://svn.science.uu.nl/repos/sci.knobb101.SubgroupDiscovery@2179 3a85a776-1598-11dc-ad04-adf9e166afdd