killMembers() is alive.
CandidateQueue calls Subgroup.killMembers() when adding Subgroups to the Queue.
Subgroup.getMemebers() will recreate would they be needed again.

Process reviveMembers() re-creates Subgroup memebrs for all Subgroups in the ResultSet.
This is done to avoid potential bugs.
When the ResultWindow is open, and the Table is altered, it might become impossible to re-create the members.
The ConditionList for the Subgroup might refer to Column that are altered.
For example, the AttributeType changed, or the missing values.



git-svn-id: https://svn.science.uu.nl/repos/sci.knobb101.SubgroupDiscovery@2248 3a85a776-1598-11dc-ad04-adf9e166afdd