Column update, itsNominals List<String> is now itsNominalz int[]. So, the data consists of indexes into the itsDistinctValues List<String>.
When itsDistinctValues was introduced, memory use went from o(8n) to o(8n + 8c) -> o(16n) bytes.
When itsDistinctValuesMap was introduced it increased to 9x / 11x the original size.
Now that the data is in int[], there is a tradeoff around (n = 18c) or (c = .055n).
This could be much lower using a special IndexMap class (n = 8c), and data[] of smaller type (n = 4.28c).
But it requires writing a special Bi-directional Map.

Added asserts for ORDINAL.

NOTE
Some of the old code was not tested, and seemed to be incorrect in some circumstances.
New code may still suffer from this.



git-svn-id: https://svn.science.uu.nl/repos/sci.knobb101.SubgroupDiscovery@2253 3a85a776-1598-11dc-ad04-adf9e166afdd