QualityMeasure update for nominal measures. ConvexHull.HullPoint coordinates are ALWAYS integers, so float replaced by int.
QualityMeasure:
Old calculate used float for the True Positive and Coverage counts.
This would cause rounding errors for large datasets.
Parameters are now int.
When required, these are converted to double, which is safe (no loss in precision).

Also, calculate now uses consistent naming for all measures, based on Head/Body.

Finally, all measures should now check for border cases of H or B==0, and NaNs.



ConvexHull.HullPoint:
HullPoint is a static inner class now, as it needs no reference to (members of)
the enclosing class.
The coordinates are integer counts (number of positives and negatives), so
instead of float they are now int.
This avoids the same issues for large datasets as QualityMeasure (see above).



ConvexHull:
See HullPoint float->int.
Internal calculations are adapted to not overflow.
Also, there should be less (no) rounding errors in the calculations.



SubgroupDiscovery + gui/ROCCurve are affected by all of the above.



git-svn-id: https://svn.science.uu.nl/repos/sci.knobb101.SubgroupDiscovery@2583 3a85a776-1598-11dc-ad04-adf9e166afdd