SubDisc: QualityMeasure.java, getROCHeaven/Hell() call calculate(QM, N, TTC, HB, B) directly.
Before, these methods called calculate(HB, B).
It is a public method that performs extra checks on externally supplied data.
These checks are not required for these methods, and would fail.
It is safe to call calculate(QM, N, TTC, HB, B) directly (also, the methods are unrelated to mining).



git-svn-id: https://svn.science.uu.nl/repos/sci.knobb101.SubgroupDiscovery@3235 3a85a776-1598-11dc-ad04-adf9e166afdd