SubDisc: SubgroupDiscovery SDMM algorithm is so fast that it creates too much contention on CandidateSet/ResultSet lock.
Quick change to mitigate the problem somewhat.
Two checks that are performed by CandidateQueue.add() are peformed by SubgroupDiscovery.checkAndLog().
When they fail, this avoids a needless synchronized statement on the itsCheckLock.
For now, CandidateQueue performs the checks again, will need to check other code, before creating asserts.

For ResultSet a dirty hasPotential() methods will be introduced.
When Subgroup.score < ResultSet.lowestScore there is no need to try to add it.
This avoids some of the locks.
It will be dirty, as multiple threads check this, and by the time ResultSet.add() is calles, lowestscore might have changed.
In this case the Subgroup is renounced after all.