SubDisc: BUGFIX SubgroupSet.postProcess() now calls CoverBasedSubgroupSelection.postProcess().
Major change: the CBSS technique as presented in the papers is fundamentally flawed.
It can not properly deal with measures that yield negative scores.
Among these are WRAcc en Numeric WRAcc, as presented in the papers.
In fact, the behaviour is the exact opposite of what it aspires to accomplish.
This change fixes the technique.

Minor change: original code contained a bug.
All (remaining) Candidates/Subgroups can have a priority/quality of NEGATIVE_INFINITY or NaN.
The former seems to be a valid result for some model classes/quality measures.
Even if not, most do not check their result.
In this case the original code would not behave correctly.
It would not select a Best Candidate/Subgroup.
But as the selection defaulted to 0, the top-scoring Candidate/Subgroup would be selected.
(Over and over again.)